+++ scm/gitlab.go
@@ -305,14 +305,7 @@ func (c Gitlab) filter(group string, ps []*gitlab.Project) []Repo {
 		// https://github.com/gabrie30/ghorg/issues/228
 		// https://github.com/gabrie30/ghorg/issues/267
 		if !gitLabAllGroups && !gitLabAllUsers {
-			// cloud instance of gitlab
-			if os.Getenv("GHORG_SCM_BASE_URL") == "" {
-				path = strings.TrimPrefix(path, group)
-			} else {
-				if os.Getenv("GHORG_OUTPUT_DIR") == "" {
-					path = strings.TrimPrefix(path, group)
-				}
-			}
+			path = strings.TrimPrefix(path, group)
 		}
 
 		r.Path = path