+++ CHANGELOG.md
@@ -14,6 +14,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 - gitlab hash concurrency issues
 - all-users command directory nesting
 - ls command to work with output dirs
+- gitlab group name fix when using --preserve-dir and --output-dir; thanks @nudgegoonies
 ### Security
 - Bump github.com/ktrysmt/go-bitbucket from 0.9.54 to 0.9.55
 - Bump github.com/xanzy/go-gitlab from 0.76.0 to 0.77.0