+++ CHANGELOG.md
@@ -5,7 +5,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 
 ## [1.9.3] - unreleased
 ### Added
-- Better examples for GitLab
+- Better examples for GitLab; thanks @nudgegoonies
 - Better tests for local gitlab enterprise
 ### Changed
 ### Deprecated
@@ -14,7 +14,6 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 - gitlab hash concurrency issues
 - all-users command directory nesting
 - ls command to work with output dirs
-- gitlab group name fix when using --preserve-dir and --output-dir; thanks @nudgegoonies
 ### Security
 - Bump github.com/ktrysmt/go-bitbucket from 0.9.54 to 0.9.55
 - Bump github.com/xanzy/go-gitlab from 0.76.0 to 0.77.0