+++ CHANGELOG.md
@@ -3,7 +3,7 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 
-## [1.9.3] - unreleased
+## [1.9.3] - 1/7/23
 ### Added
 - Better examples for GitLab
 - Better tests for local gitlab enterprise